Fixed 'assertion failed' error that occurred with -p flag and read pairs if no alignment was found for the mate with the higher number of k-mer word hits in the hash index. Added info about read number and name to error message when smalt check task failed.

+++ misc/simread.c
@@ -803,7 +803,7 @@ int main(int argc, char *argv[])
   //RANSEED(RANDOM_SEED);
   RANSEED(randseed);
 
-  basnum = readlen*readnum;
+  basnum = ((uint64_t) readlen)*readnum;
   if (insertsiz != 0)
     basnum *= 2;
   printf("Simulate %llu bases ...\n", (LLUINT) basnum);
